+++ lib/src/components/components.dart
@@ -36,6 +36,7 @@ export 'flowbite_modal.dart';
 export 'flowbite_outlined_button.dart';
 export 'flowbite_progress_bar.dart';
 export 'flowbite_radio_input.dart';
+export 'flowbite_radio_input_card.dart';
 export 'flowbite_range_slider.dart';
 export 'flowbite_sidebar.dart';
 export 'flowbite_spinner.dart';